Refactor to create an abstract StorageService extended to obtain AWSS3Service and AWSMinioService
