+++ avni-server-api/src/main/resources/application.properties
@@ -54,7 +54,6 @@ server.compression.min-response-size=2048
 server.compression.mime-types=application/json,application/xml,text/html,text/xml,text/plain,application/hal+json,application/javascript
 
 # S3 Media
-aws.s3.url=${OPENCHS_S3_URL:https://s3.ap-south-1.amazonaws.com}
 aws.accessKeyId=${OPENCHS_IAM_USER_ACCESS_KEY:dummy}
 aws.secretAccessKey=${OPENCHS_IAM_USER_SECRET_ACCESS_KEY:dummy}
 avni.bucketName=${OPENCHS_BUCKET_NAME:dummy}
